[T] Move autocheck generators in own header file.
Replace makeOrderOfEps with special generator.
With the new generator there is no need to discard
values.
